Changed CPU from milliseconds to microseconds
Due to incorrect documentation, CPU was captured in inconsistent ways between sqlTrace and Extended Events.
CPU is now collected only as microseconds internally and persisted as milliseconds in the analysis database, microseconds everywhere else

+++ WorkloadTools/ExecutionWorkloadEvent.cs
@@ -16,7 +16,7 @@ namespace WorkloadTools
         public string HostName { get; set; }
         public long? Reads { get; set; }
         public long? Writes { get; set; }
-        public int? CPU { get; set; }
+        public long? CPU { get; set; }
         public long? Duration { get; set; }
 
     }
